Added the folder made by me
The score.txt file already contains some data that were my test runs.
NOTE: The character scanning is working only when I wrote my first name only.  My first name is 5 characters long. Please, see into the issue if you can.
But the rest is all good!
The file is being written and all and even showing the scores when called upon.
Please, upload the edited or the final program so that I can write the scripts for the three of us. Please try to do it before 10 tomorrow morning.
Thanks, Brother! We got this! Chill hai ek dam!